Completely ported to nextjs - Deploy
Disabled image optimization in next config which means the build is using an experimental feature. Will investigate risks.

Interestingly, get static props only works in the base page component and not any children components. I had planned to only get data on build but unless I change the component structure of featured_projects and  about_skills it will stay the way it is.